refactor(tools): consolidate campaign tools with shared utilities
- Create src/tools/utils.ts with common tool patterns
- Standardize JWT handling with commonSchemas
- Replace repetitive error/success responses with utility functions
- Consolidate authentication headers and context detection
- Reduce campaign.ts from ~430 to ~280 lines (35% reduction)
- Improve maintainability and type safety across all campaign tools
