+++ src/hooks/useOpenAIKey.ts
@@ -49,7 +49,6 @@ export function useOpenAIKey(): UseOpenAIKeyReturn {
     {
       onSuccess: (hasKey) => setHasApiKey(hasKey),
       onError: () => setHasApiKey(false),
-      autoExecute: true,
     }
   );
 
@@ -89,7 +88,7 @@ export function useOpenAIKey(): UseOpenAIKeyReturn {
 
   useEffect(() => {
     checkApiKeyStatus.execute();
-  }, [checkApiKeyStatus.execute]);
+  }, [checkApiKeyStatus.execute]); // Only run once on mount
 
   return {
     hasApiKey,
