+++ src/tools/pdf/index.ts
@@ -1,0 +1,28 @@
+// Import all PDF tools
+
+import { getPdfStats, listPdfFiles } from "./list-tools";
+import { updatePdfMetadata } from "./metadata-tools";
+import {
+  generatePdfUploadUrl,
+  ingestPdfFile,
+  uploadPdfFile,
+} from "./upload-tools";
+
+export { getPdfStats, listPdfFiles } from "./list-tools";
+export { updatePdfMetadata } from "./metadata-tools";
+// Export all PDF tools
+export {
+  generatePdfUploadUrl,
+  ingestPdfFile,
+  uploadPdfFile,
+} from "./upload-tools";
+
+// Export the tools object for backward compatibility
+export const pdfTools = {
+  uploadPdfFile,
+  ingestPdfFile,
+  generatePdfUploadUrl,
+  listPdfFiles,
+  getPdfStats,
+  updatePdfMetadata,
+};
